requirements.txt
django==3.1
djangorestframework
django-sslserver
drf_yasg